adding public and private routes via verifyToken

+++ userHandlers.js
@@ -80,7 +80,6 @@ const postUser = (req, res) => {
     });
 };
 
-
 const getUserByEmailWithPasswordAndPassToNext = (req, res, next) => {
   const { email } = req.body;
   database
@@ -88,7 +87,7 @@ const getUserByEmailWithPasswordAndPassToNext = (req, res, next) => {
     .then(([users]) => {
       if (users[0] != null) {
         req.user = users[0];
-        next()
+        next();
       } else {
         res.sendStatus(401);
       }
@@ -102,46 +101,51 @@ const getUserByEmailWithPasswordAndPassToNext = (req, res, next) => {
 const updateUser = (req, res) => {
   const id = parseInt(req.params.id);
   const { firstname, lastname, email, city, language } = req.body;
-  console.log(req.body);
+  if (req.payload.sub !== id) {
+    res.sendStatus(403);
+  } else {
+    database
+      .query(
+        "update users set firstname = ?, lastname = ?, email = ?, city = ?, language = ? where id = ?",
+        [firstname, lastname, email, city, language, id]
+      )
 
-  database
-    .query(
-      "update users set firstname = ?, lastname = ?, email = ?, city = ?, language = ? where id = ?",
-      [firstname, lastname, email, city, language, id]
-    )
+      .then(([result]) => {
+        if (result.affectedRows === 0) {
+          res.status(404).send("Ressource not found"); // status solo = que code erreur + mesasge avec send sinon sans message sendStatus
+        } else {
+          res.sendStatus(204);
+        }
+      })
 
-    .then(([result]) => {
-      if (result.affectedRows === 0) {
-        res.status(404).send("Ressource not found");
-      } else {
-        res.sendStatus(204);
-      }
-    })
-
-    .catch((err) => {
-      console.error(err);
-      res.status(500).send("Error editing the user");
-    });
+      .catch((err) => {
+        console.error(err);
+        res.status(500).send("Error editing the user");
+      });
+  }
 };
 
 const deleteUser = (req, res) => {
   const id = parseInt(req.params.id);
+  if (req.payload.sub !== id) {
+    res.sendStatus(403);
+  } else {
+    database
+      .query("delete from users where id = ?", [id])
 
-  database
-    .query("delete from users where id = ?", [id])
+      .then(([result]) => {
+        if (result.affectedRows === 0) {
+          res.status(404).send("User not Found");
+        } else {
+          res.sendStatus(204);
+        }
+      })
 
-    .then(([result]) => {
-      if (result.affectedRows === 0) {
-        res.status(404).send("User not Found");
-      } else {
-        res.sendStatus(204);
-      }
-    })
-
-    .catch((err) => {
-      console.error(err);
-      res.status(500).send("Error deleting user");
-    });
+      .catch((err) => {
+        console.error(err);
+        res.status(500).send("Error deleting user");
+      });
+  }
 };
 
 module.exports = {
@@ -150,5 +154,5 @@ module.exports = {
   postUser,
   updateUser,
   deleteUser,
-  getUserByEmailWithPasswordAndPassToNext
+  getUserByEmailWithPasswordAndPassToNext,
 };
